updated channel page and list
