+++ mobile/src/components/features/channels/AddChannel.tsx
@@ -1,0 +1,8 @@
+
+type Props = {}
+
+export const AddChannel = (props: Props) => {
+    return (
+        <div>AddChannel</div>
+    )
+}
