+++ mobile/src/components/features/channels/index.ts
@@ -1,1 +1,3 @@
-export { Channels } from './Channels'
+export { ChannelList } from './ChannelList'
+export { AddChannel } from './AddChannel'
+export { AddChannelMembers } from './AddChannelMembers'
