--- mobile/src/pages/channels/ChannelList.tsx
@@ -1,37 +1,0 @@
-import { IonContent, IonHeader, IonPage, IonTitle, IonToolbar, IonItem, IonLabel } from '@ionic/react'
-import { IoAdd } from 'react-icons/io5'
-import { Channels } from '../../components/features/channels';
-
-type Props = {}
-
-export const ChannelList = (props: Props) => {
-
-    return (
-        <IonPage>
-            <IonHeader translucent>
-                <IonToolbar>
-                    <IonTitle>Raven</IonTitle>
-                </IonToolbar>
-            </IonHeader>
-            <IonContent fullscreen>
-                <IonHeader collapse="condense" translucent>
-                    <IonToolbar>
-                        <IonTitle size="large">Raven</IonTitle>
-                    </IonToolbar>
-                </IonHeader>
-                <div className='ion-padding-horizontal py-2'>
-                    <IonLabel className='text-sm font-medium' color='medium'>Channels</IonLabel>
-                </div>
-                <Channels />
-                <IonItem lines='none'>
-                    <div slot='start'>
-                        <IoAdd size='24' color='var(--ion-color-medium)' />
-                    </div>
-                    <IonLabel color='medium'>
-                        Add Channel
-                    </IonLabel>
-                </IonItem>
-            </IonContent>
-        </IonPage>
-    )
-}
