+++ mobile/src/pages/channels/index.ts
@@ -1,1 +1,1 @@
-export { ChannelList } from './ChannelList'
+export { Channels } from './Channels'
